creation de la classe Retrait + ...

+++ Back-end/src/main/java/org/enchere/backend/Repository/ArticleVenduRepository.java
@@ -3,5 +3,5 @@ package org.enchere.backend.Repository;
 import org.enchere.backend.model.ArticleVendu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleVenduRepository extends JpaRepository<ArticleVendu, Long> {
+public interface ArticleVenduRepository extends JpaRepository<ArticleVendu, Integer> {
 }
